client/requirements.txt
psycopg2-binary==2.9.11
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.1
fake-useragent>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0